+++ throttle/src/CMakeLists.txt
@@ -2,4 +2,4 @@ add_subdirectory(guards)
 
 target_sources(app PRIVATE main.cpp)
 target_sources(app PRIVATE server.cpp)
-target_sources(app PRIVATE ThrottleValve.cpp)
+target_sources(app PRIVATE throttle_valve.cpp)
